+++ src/transformer/utils.ts
@@ -181,6 +181,8 @@ export function expressionForPropertyName(propName : ts.PropertyName) {
         return propName.expression; // TODO: reuse of node might not be a good idea, but it does work
     } else if (ts.isIdentifier(propName)) {
         return ts.factory.createStringLiteral(propName.text);
+    } else if (ts.isStringLiteral(propName)) {
+        return ts.factory.createStringLiteral(propName.text);
     } else {
         throw new Error(`Unexpected property name node of type '${ts.SyntaxKind[propName.kind]}'! Please file a bug!`);
     }