+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # vNext
 - Fix for compile-time crash when mapped type has no `aliasTypeArguments`
-
+- Fix for properties which use a string literal name
 
 # v0.4.14
 - Support for metadata on properties with computed names (including symbols)