+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# vNext
+# v0.4.15
 - Fix for compile-time crash when mapped type has no `aliasTypeArguments`
 - Fix for properties which use a string literal name
 - Add missing support for class expressions